+++ EsportsTour/Controllers/JeuxController.cs
@@ -13,7 +13,7 @@ using Projet.Net.Models;
 
 namespace EsportsTour.Controllers
 {
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     public class JeuxController : Controller
     {
         private readonly EsportsDbContext _context;
@@ -50,6 +50,7 @@ namespace EsportsTour.Controllers
         }
 
         // GET: Jeux/Create
+        [Authorize(Roles = "Admin")]
         public IActionResult Create()
         {
             return View();
@@ -60,6 +61,7 @@ namespace EsportsTour.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Create([Bind("Id,NomJeu,Categorie,ImgJeu")] Jeux jeux)
         {
             if (ModelState.IsValid)
@@ -77,6 +79,7 @@ namespace EsportsTour.Controllers
         }
 
         // GET: Jeux/Edit/5
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Jeux == null)
@@ -97,6 +100,7 @@ namespace EsportsTour.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Edit(int id, [Bind("Id,NomJeu,Categorie,ImgJeu")] Jeux jeux)
         {
             if (id != jeux.Id)
@@ -128,6 +132,7 @@ namespace EsportsTour.Controllers
         }
 
         // GET: Jeux/Delete/5
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null || _context.Jeux == null)
@@ -148,6 +153,7 @@ namespace EsportsTour.Controllers
         // POST: Jeux/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             if (_context.Jeux == null)
@@ -155,18 +161,26 @@ namespace EsportsTour.Controllers
                 return Problem("Entity set 'EsportsDbContext.Jeux'  is null.");
             }
             var jeux = await _context.Jeux.FindAsync(id);
+            bool hasResultats = _context.Tournois.Any(r => r.JeuId == id);
+
+            if (hasResultats)
+            {
+                ModelState.AddModelError(string.Empty, "Impossible de supprimer cette Jeux car il y a des Tournoi associés.");
+                return View(jeux);
+            }
             if (jeux != null)
             {
                 _context.Jeux.Remove(jeux);
             }
-            
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool JeuxExists(int id)
         {
-          return (_context.Jeux?.Any(e => e.Id == id)).GetValueOrDefault();
+            return (_context.Jeux?.Any(e => e.Id == id)).GetValueOrDefault();
         }
+
     }
 }
